+++ packages/app-builder/src/root.tsx
@@ -12,8 +12,6 @@ import {
   Scripts,
   ScrollRestoration,
   useLoaderData,
-  useLocation,
-  useMatches,
   useRouteError,
 } from '@remix-run/react';
 import { captureRemixErrorBoundaryError, withSentry } from '@sentry/remix';
@@ -23,7 +21,6 @@ import { useTranslation } from 'react-i18next';
 import {
   AuthenticityTokenProvider,
   createAuthenticityToken,
-  useHydrated,
 } from 'remix-utils';
 import { Tooltip } from 'ui-design-system';
 import { LogoStandard } from 'ui-icons';
@@ -31,9 +28,9 @@ import { LogoStandard } from 'ui-icons';
 import { ErrorComponent } from './components/ErrorComponent';
 import { getToastMessage, MarbleToaster } from './components/MarbleToaster';
 import { serverServices } from './services/init.server';
+import { SegmentScript } from './services/segment';
 import tailwindStyles from './tailwind.css';
 import { getClientEnvVars } from './utils/environment.server';
-import getPageViewNameAndProps from './utils/getPageviewNameAndProps';
 
 export const links: LinksFunction = () => [
   { rel: 'stylesheet', href: tailwindStyles },
@@ -151,31 +148,12 @@ function App() {
     void i18n.changeLanguage(locale);
   }, [locale, i18n]);
 
-  const location = useLocation();
-  const isHydrated = useHydrated();
-  const matches = useMatches();
-  const thisPage = matches[matches.length - 1];
-  useEffect(() => {
-    if (isHydrated) {
-      const { name, properties } = getPageViewNameAndProps(thisPage);
-      window.analytics.page(name, properties);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [location, thisPage.id, isHydrated]);
-
   return (
     <html lang={locale} dir={i18n.dir()}>
       <head>
         <Meta />
         <Links />
-        <script
-          dangerouslySetInnerHTML={{
-            __html: `
-        !function(){var analytics=window.analytics=window.analytics||[];if(!analytics.initialize)if(analytics.invoked)window.console&&console.error&&console.error("Segment snippet included twice.");else{analytics.invoked=!0;analytics.methods=["trackSubmit","trackClick","trackLink","trackForm","pageview","identify","reset","group","track","ready","alias","debug","page","once","off","on","addSourceMiddleware","addIntegrationMiddleware","setAnonymousId","addDestinationMiddleware"];analytics.factory=function(e){return function(){if(window.analytics.initialized)return window.analytics[e].apply(window.analytics,arguments);var i=Array.prototype.slice.call(arguments);i.unshift(e);analytics.push(i);return analytics}};for(var i=0;i<analytics.methods.length;i++){var key=analytics.methods[i];analytics[key]=analytics.factory(key)}analytics.load=function(key,i){var t=document.createElement("script");t.type="text/javascript";t.async=!0;t.src="https://cdn.segment.com/analytics.js/v1/" + key + "/analytics.min.js";var n=document.getElementsByTagName("script")[0];n.parentNode.insertBefore(t,n);analytics._loadOptions=i};analytics._writeKey="${ENV.SEGMENT_WRITE_KEY}";;analytics.SNIPPET_VERSION="4.16.1";
-        analytics.load("${ENV.SEGMENT_WRITE_KEY}");
-        }}();`,
-          }}
-        />
+        <SegmentScript key={ENV.SEGMENT_WRITE_KEY} />
       </head>
       <body className="selection:text-grey-00 h-screen w-full overflow-hidden antialiased selection:bg-purple-100">
         <AuthenticityTokenProvider token={csrf}>
