+++ packages/app-builder/src/services/segment/segment.d.ts
@@ -1,5 +1,3 @@
-export {};
-
 declare global {
   interface Window {
     analytics: SegmentAnalytics.AnalyticsJS;
